impl ExceptionHandling Methodlogy
impl ExceptionHandling Methodlogy

+++ section2/accounts/src/main/java/com/eazybytes/accounts/entity/Accounts.java
@@ -12,7 +12,7 @@ public class Accounts extends BaseEntity {
 
     @Id
     @Column(name = "account_number")
-    public Integer accountNumber;
+    public long accountNumber;
 
     @Column(name = "customer_id")
     public long customerID;
